pypypypy/requirements.txt
fastapi
uvicorn
asyncpg
pyjwt 